pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iotcli"
version = "2.0.0"
description = "Universal IoT device control CLI — for humans and AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "IoTviaAI" }]
keywords = ["iot", "smart-home", "cli", "ai-agent", "tuya", "xiaomi", "mqtt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "paho-mqtt>=1.6.1",
    "cryptography>=41.0.0",
    "rich>=13.0.0",
    "InquirerPy>=0.3.4",
    "tabulate>=0.9.0",
    "aiohttp>=3.9.0",
    "tinytuya>=1.13.0",
    "netifaces>=0.11.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[project.scripts]
iotcli = "iotcli.cli.app:main"

[project.urls]
Homepage = "https://github.com/iotviaai/iotcli"
Issues = "https://github.com/iotviaai/iotcli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/iotcli"]

[tool.ruff]
target-version = "py310"
line-length = 100
